refactor: extract date-formatting utilities into DateFormatUtils.ts (#590)
Moves formatTableDate(), formatDiscordTimestamp(), and formatPlaytimeHours()
out of profile.command.ts into src/functions/DateFormatUtils.ts, and removes
two divergent local reimplementations from todo.command.ts and
avatar-history.command.ts. Updates all 14 import sites accordingly.
