+++ src/commands/game-completion/completion-autocomplete.utils.ts
@@ -5,7 +5,7 @@ import { sanitizeUserInput } from "../../functions/InteractionUtils.js";
 import { formatGameTitleWithYear } from "../../functions/GameTitleAutocompleteUtils.js";
 import Game, { type IPlatformDef } from "../../classes/Game.js";
 import Member from "../../classes/Member.js";
-import { formatTableDate } from "../profile.command.js";
+import { formatTableDate } from "../../functions/DateFormatUtils.js";
 import { STANDARD_PLATFORM_IDS } from "../../config/standardPlatforms.js";
 import { isPositiveInt } from "../../utilities/ValidationUtils.js";
